# pkg/bgpv1/annotations: Optimize annotations Errors