Avoiding to use  fmt.Sprintf() so that Go won't over-allocate the memory.

issue reference : cilium#19571

Result:

```
$ go test -v -benchmem -run=^$ -bench 'BenchmarkErrNotVRouterAnnoError' github.com/cilium/cilium/pkg/bgpv1 > BenchmarkErrNotVRouterAnnoError_old.txt
$ go test -v -benchmem -run=^$ -bench 'BenchmarkErrErrNoASNAnno' github.com/cilium/cilium/pkg/bgpv1 > BenchmarkErrErrNoASNAnno_old.txt
$ go test -v -benchmem -run=^$ -bench 'BenchmarkErrASNAnno' github.com/cilium/cilium/pkg/bgpv1 > BenchmarkErrASNAnno_old.txt
$ go test -v -benchmem -run=^$ -bench 'BenchmarkErrNotVRouterAnnoError' github.com/cilium/cilium/pkg/bgpv1 > BenchmarkErrNotVRouterAnnoError_new.txt
$ go test -v -benchmem -run=^$ -bench 'BenchmarkErrErrNoASNAnno' github.com/cilium/cilium/pkg/bgpv1 > BenchmarkErrErrNoASNAnno_new.txt
$ go test -v -benchmem -run=^$ -bench 'BenchmarkErrASNAnno' github.com/cilium/cilium/pkg/bgpv1 > BenchmarkErrASNAnno_new.txt
$ benchcmp BenchmarkErrNotVRouterAnnoError_old.txt BenchmarkErrNotVRouterAnnoError_new.txt
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                              old ns/op     new ns/op     delta
BenchmarkErrNotVRouterAnnoError-12     186           47.2          -74.67%

benchmark                              old allocs     new allocs     delta
BenchmarkErrNotVRouterAnnoError-12     2              1              -50.00%

benchmark                              old bytes     new bytes     delta
BenchmarkErrNotVRouterAnnoError-12     96            80            -16.67%
$ benchcmp BenchmarkErrErrNoASNAnno_old.txt BenchmarkErrErrNoASNAnno_new.txt
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                       old ns/op     new ns/op     delta
BenchmarkErrErrNoASNAnno-12     143           45.0          -68.49%

benchmark                       old allocs     new allocs     delta
BenchmarkErrErrNoASNAnno-12     2              1              -50.00%

benchmark                       old bytes     new bytes     delta
$ benchcmp BenchmarkErrASNAnno_old.txt BenchmarkErrASNAnno_new.txt
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                  old ns/op     new ns/op     delta
BenchmarkErrASNAnno-12     252           61.6          -75.59%

benchmark                  old allocs     new allocs     delta
BenchmarkErrASNAnno-12     3              1              -66.67%

benchmark                  old bytes     new bytes     delta
BenchmarkErrASNAnno-12     112           80            -28.57%
```

Signed-off-by: MikeLing <sabergeass@gmail.com>